Rsync option

+++ lib/librarian/puppet/util.rb
@@ -1,3 +1,5 @@
+require 'rsync'
+
 module Librarian
   module Puppet
 
@@ -13,16 +15,26 @@ module Librarian
         environment.logger.warn(*args, &block)
       end
 
+      def rsync?
+          environment.config_db.local['rsync'] == 'true'
+      end
+
       # workaround Issue #173 FileUtils.cp_r will fail if there is a symlink that points to a missing file
       # or when the symlink is copied before the target file when preserve is true
       # see also https://tickets.opscode.com/browse/CHEF-833
+      #
+      # If the rsync configuration parameter is set, use rsync instead of FileUtils
       def cp_r(src, dest)
-        begin
-          FileUtils.cp_r(src, dest, :preserve => true)
-        rescue Errno::ENOENT
-          debug { "Failed to copy from #{src} to #{dest} preserving file types, trying again without preserving them" }
-          FileUtils.rm_rf(dest)
-          FileUtils.cp_r(src, dest)
+        if rsync?
+          Rsync.run(File.join(src, "/"), dest, ['-avz', '--delete'])
+        else
+          begin
+            FileUtils.cp_r(src, dest, :preserve => true)
+          rescue Errno::ENOENT
+            debug { "Failed to copy from #{src} to #{dest} preserving file types, trying again without preserving them" }
+            FileUtils.rm_rf(dest)
+            FileUtils.cp_r(src, dest)
+          end
         end
       end
